created some more common methods
